Python/animeYTB-main/requirements.txt
pillow==9.5.0
moviepy==1.0.3
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
firebase-admin
requests